requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
anytree==2.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arm_pytorch_utilities==0.4.3
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
cloudpickle==3.1.1
cmeel==0.57.1
cmeel-assimp==5.4.3.1
cmeel-boost==1.83.0
cmeel-console-bridge==1.0.2.2
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml==2.6.2.3
cmeel-urdfdom==3.1.1.1
cmeel-zlib==1.3.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
dash==3.0.1
datasets==3.4.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
eigenpy==3.5.1
embreex==2.17.7.post6
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fast_kinematics==0.2.2
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.6.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.8.0
grpcio==1.71.0
gymnasium==0.29.1
h11==0.14.0
h5py==3.13.0
hpp-fcl==2.4.4
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
ikpy==3.4.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.2
jaxlib==0.5.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.3.1
mani_skill==3.0.0b20
manifold3d==3.0.1
mapbox_earcut==1.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.2
ml_dtypes==0.5.1
mplib==0.1.1
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.5
nlopt==2.7.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pin==2.7.0
platformdirs==4.3.6
plotly==6.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycollada==0.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyglet==2.1.3
Pygments==2.19.1
pynvml==12.0.0
PyOpenGL==3.1.0
pyparsing==3.2.1
pyperclip==1.9.0
pyquaternion==0.9.9
pyrender==0.1.45
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-kinematics==0.7.5
pytorch-seed==0.2.0
pytransform3d==3.14.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
rsl-rl-lib==2.3.0
rtree==1.4.0
sapien==3.0.0b1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.23.1
setproctitle==1.3.5
shapely==2.0.7
shtab==1.7.1
six==1.17.0
skrl==1.4.2
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
svg.path==6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
toppra==0.6.3
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.6.4
triton==3.2.0
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2025.1
urdfpy==0.0.22
uri-template==1.3.0
urllib3==2.3.0
vhacdx==0.0.8.post2
wandb==0.19.8
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xatlas==0.0.9
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
